Move things into HotspotDetails
